requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
GitPython==3.1.40
aiofiles==23.2.1
requests==2.31.0